metacat/client/requirements.txt
annotated-types
certifi
charset-normalizer
idna
pydantic
pydantic_core
requests
requests-oauthlib
typing_extensions
urllib3
python-dotenv
